fix error, now open 10gb file (but before this slow indexating it

+++ Rope/ITextBuffer.cs
@@ -49,7 +49,7 @@ namespace TextBuffer
     {
         public char this[long index] { get; }
 
-        public int Length { get; }
+        public long Length { get; }
 
         public long LengthEx(IntPtr state);
